+++ src/main/java/zoo/pubg/domain/Match.java
@@ -3,10 +3,14 @@ package zoo.pubg.domain;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import java.time.LocalDateTime;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
+@AllArgsConstructor
+@NoArgsConstructor
 public class Match {
 
     @Id
