+++ src/main/java/zoo/pubg/domain/RosterMatchResult.java
@@ -1,0 +1,21 @@
+package zoo.pubg.domain;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+
+@Entity
+public class RosterMatchResult {
+
+    @Id
+    private String rosterId;
+
+    private int rank;
+    private int teamId;
+    private String playerId;
+
+    @ManyToOne
+    @JoinColumn(name = "matchId", referencedColumnName = "matchId")
+    private Match match;
+}
