--- src/main/java/zoo/pubg/service/dto/match/data/Data.java
@@ -1,10 +1,0 @@
-package zoo.pubg.service.dto.match.data;
-
-import lombok.Getter;
-
-@Getter
-public class Data {
-
-    private String id;
-    private MatchAttributes attributes;
-}
