--- src/main/java/zoo/pubg/service/dto/match/included/asset/AssetAttributes.java
@@ -1,10 +1,0 @@
-package zoo.pubg.service.dto.match.included.asset;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import java.time.ZonedDateTime;
-
-public record AssetAttributes(String URL,
-                              @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ssXXX")
-                              ZonedDateTime createdAt) {
-
-}
