--- src/main/java/zoo/pubg/service/dto/match/included/participant/HistoryAttributes.java
@@ -1,5 +1,0 @@
-package zoo.pubg.service.dto.match.included.participant;
-
-public record HistoryAttributes(PlayerMatchStats stats) {
-
-}
