--- src/main/java/zoo/pubg/service/dto/match/included/participant/PlayerMatchStats.java
@@ -1,6 +1,0 @@
-package zoo.pubg.service.dto.match.included.participant;
-
-public record PlayerMatchStats(int dbno, int kills, int assists, int headshotKills, float damageDealt,
-                               float longestKill) {
-
-}
