--- src/main/java/zoo/pubg/service/dto/match/included/roster/RosterAttributes.java
@@ -1,7 +1,0 @@
-package zoo.pubg.service.dto.match.included.roster;
-
-import zoo.pubg.service.dto.match.included.roster.attributes.RosterMatchStats;
-
-public record RosterAttributes(RosterMatchStats stats) {
-
-}
