--- src/main/java/zoo/pubg/service/dto/match/included/roster/RosterRelationships.java
@@ -1,8 +1,0 @@
-package zoo.pubg.service.dto.match.included.roster;
-
-import zoo.pubg.service.dto.match.included.roster.relationships.RosterParticipants;
-
-public record RosterRelationships(
-        RosterParticipants participants) {
-
-}
