--- src/main/java/zoo/pubg/service/dto/match/included/roster/attributes/RosterMatchStats.java
@@ -1,4 +1,0 @@
-package zoo.pubg.service.dto.match.included.roster.attributes;
-
-public record RosterMatchStats(int rank, int teamId) {
-}
