--- src/main/java/zoo/pubg/service/dto/match/included/roster/relationships/RosterParticipants.java
@@ -1,8 +1,0 @@
-package zoo.pubg.service.dto.match.included.roster.relationships;
-
-import java.util.List;
-import zoo.pubg.service.dto.match.included.roster.relationships.participants.RosterParticipantData;
-
-public record RosterParticipants(List<RosterParticipantData> data) {
-
-}
