--- src/main/java/zoo/pubg/service/dto/match/included/roster/relationships/participants/RosterParticipantData.java
@@ -1,5 +1,0 @@
-package zoo.pubg.service.dto.match.included.roster.relationships.participants;
-
-public record RosterParticipantData(String id) {
-
-}
