--- src/main/java/zoo/pubg/service/dto/player/MatchData.java
@@ -1,4 +1,0 @@
-package zoo.pubg.service.dto.player;
-
-public record MatchData(String type, String id) {
-}
