--- src/main/java/zoo/pubg/service/dto/player/Matches.java
@@ -1,9 +1,0 @@
-package zoo.pubg.service.dto.player;
-
-import java.util.List;
-
-public record Matches(List<MatchData> data) {
-    public List<String> getMatchIds() {
-        return data.stream().map(MatchData::id).toList();
-    }
-}
