--- src/main/java/zoo/pubg/service/dto/player/PlayerAttributes.java
@@ -1,4 +1,0 @@
-package zoo.pubg.service.dto.player;
-
-public record PlayerAttributes(String shardId, String clanId, String name) {
-}
