--- src/main/java/zoo/pubg/service/dto/player/PlayerDto.java
@@ -1,6 +1,0 @@
-package zoo.pubg.service.dto.player;
-
-import java.util.List;
-
-public record PlayerDto(List<PlayerData> data) {
-}
