--- src/main/java/zoo/pubg/service/dto/player/PlayerRelationships.java
@@ -1,9 +1,0 @@
-package zoo.pubg.service.dto.player;
-
-import java.util.List;
-
-public record PlayerRelationships(Matches matches) {
-    public List<String> getMatchIds() {
-        return matches.getMatchIds();
-    }
-}
