+++ src/main/java/zoo/pubg/service/parser/deserialization/match/data/Data.java
@@ -1,0 +1,10 @@
+package zoo.pubg.service.parser.deserialization.match.data;
+
+import lombok.Getter;
+
+@Getter
+public class Data {
+
+    private String id;
+    private MatchAttributes attributes;
+}
