+++ src/main/java/zoo/pubg/service/parser/deserialization/match/included/Participant.java
@@ -1,0 +1,18 @@
+package zoo.pubg.service.parser.deserialization.match.included;
+
+import lombok.Getter;
+import lombok.Setter;
+import zoo.pubg.domain.PlayerMatchResult;
+import zoo.pubg.service.parser.deserialization.match.included.participant.HistoryAttributes;
+
+@Getter
+@Setter
+public class Participant extends Included {
+
+    private String id;
+    private HistoryAttributes attributes;
+
+    public PlayerMatchResult toEntity() {
+        return new PlayerMatchResult();
+    }
+}
