+++ src/main/java/zoo/pubg/service/parser/deserialization/match/included/Roster.java
@@ -1,0 +1,15 @@
+package zoo.pubg.service.parser.deserialization.match.included;
+
+import lombok.Getter;
+import lombok.Setter;
+import zoo.pubg.service.parser.deserialization.match.included.roster.RosterAttributes;
+import zoo.pubg.service.parser.deserialization.match.included.roster.RosterRelationships;
+
+@Getter
+@Setter
+public class Roster extends Included {
+
+    private String id;
+    private RosterAttributes attributes;
+    private RosterRelationships relationships;
+}
