+++ src/main/java/zoo/pubg/service/parser/deserialization/match/included/participant/HistoryAttributes.java
@@ -1,0 +1,5 @@
+package zoo.pubg.service.parser.deserialization.match.included.participant;
+
+public record HistoryAttributes(PlayerMatchStats stats) {
+
+}
