+++ src/main/java/zoo/pubg/service/parser/deserialization/match/included/participant/PlayerMatchStats.java
@@ -1,0 +1,6 @@
+package zoo.pubg.service.parser.deserialization.match.included.participant;
+
+public record PlayerMatchStats(int dbno, int kills, int assists, int headshotKills, float damageDealt,
+                               float longestKill) {
+
+}
