+++ src/main/java/zoo/pubg/service/parser/deserialization/match/included/roster/RosterAttributes.java
@@ -1,0 +1,7 @@
+package zoo.pubg.service.parser.deserialization.match.included.roster;
+
+import zoo.pubg.service.parser.deserialization.match.included.roster.attributes.RosterMatchStats;
+
+public record RosterAttributes(RosterMatchStats stats) {
+
+}
