+++ src/main/java/zoo/pubg/service/parser/deserialization/match/included/roster/RosterRelationships.java
@@ -1,0 +1,8 @@
+package zoo.pubg.service.parser.deserialization.match.included.roster;
+
+import zoo.pubg.service.parser.deserialization.match.included.roster.relationships.RosterParticipants;
+
+public record RosterRelationships(
+        RosterParticipants participants) {
+
+}
