+++ src/main/java/zoo/pubg/service/parser/deserialization/match/included/roster/attributes/RosterMatchStats.java
@@ -1,0 +1,4 @@
+package zoo.pubg.service.parser.deserialization.match.included.roster.attributes;
+
+public record RosterMatchStats(int rank, int teamId) {
+}
