+++ src/main/java/zoo/pubg/service/parser/deserialization/match/included/roster/relationships/RosterParticipants.java
@@ -1,0 +1,8 @@
+package zoo.pubg.service.parser.deserialization.match.included.roster.relationships;
+
+import java.util.List;
+import zoo.pubg.service.parser.deserialization.match.included.roster.relationships.participants.RosterParticipantData;
+
+public record RosterParticipants(List<RosterParticipantData> data) {
+
+}
