+++ src/main/java/zoo/pubg/service/parser/deserialization/match/included/roster/relationships/participants/RosterParticipantData.java
@@ -1,0 +1,5 @@
+package zoo.pubg.service.parser.deserialization.match.included.roster.relationships.participants;
+
+public record RosterParticipantData(String id) {
+
+}
