+++ src/main/java/zoo/pubg/service/parser/deserialization/player/MatchData.java
@@ -1,0 +1,4 @@
+package zoo.pubg.service.parser.deserialization.player;
+
+public record MatchData(String type, String id) {
+}
