+++ src/main/java/zoo/pubg/service/parser/deserialization/player/Matches.java
@@ -1,0 +1,9 @@
+package zoo.pubg.service.parser.deserialization.player;
+
+import java.util.List;
+
+public record Matches(List<MatchData> data) {
+    public List<String> getMatchIds() {
+        return data.stream().map(MatchData::id).toList();
+    }
+}
