+++ src/main/java/zoo/pubg/service/parser/deserialization/player/PlayerAttributes.java
@@ -1,0 +1,4 @@
+package zoo.pubg.service.parser.deserialization.player;
+
+public record PlayerAttributes(String shardId, String clanId, String name) {
+}
