+++ src/main/java/zoo/pubg/service/parser/deserialization/player/PlayerDto.java
@@ -1,0 +1,6 @@
+package zoo.pubg.service.parser.deserialization.player;
+
+import java.util.List;
+
+public record PlayerDto(List<PlayerData> data) {
+}
