+++ src/main/java/zoo/pubg/service/parser/deserialization/player/PlayerRelationships.java
@@ -1,0 +1,9 @@
+package zoo.pubg.service.parser.deserialization.player;
+
+import java.util.List;
+
+public record PlayerRelationships(Matches matches) {
+    public List<String> getMatchIds() {
+        return matches.getMatchIds();
+    }
+}
