+++ src/test/java/zoo/pubg/service/parser/MatchApiParserTest.java
@@ -4,7 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import org.junit.jupiter.api.Test;
-import zoo.pubg.service.dto.match.MatchInformation;
+import zoo.pubg.domain.Match;
+import zoo.pubg.service.parser.deserialization.match.MatchInformation;
 
 class MatchApiParserTest {
 
@@ -146,15 +147,16 @@ class MatchApiParserTest {
     @Test
     void parseTest() throws JsonProcessingException {
         MatchInformation matchInformation = parser.parse(json);
+        Match match = matchInformation.getMatchEntity();
 
-        String matchId = matchInformation.getData().id(); //    b79c05af-bfb3-4744-b9d2-9ccdbe4c0746
-        String mapName = matchInformation.getData().attributes().getMapName(); // Baltic_Main
-
-        int includesSize = matchInformation.getIncluded().size();
+        String matchId = match.getMatchId(); //    b79c05af-bfb3-4744-b9d2-9ccdbe4c0746
+        String mapName = match.getMapName(); // Baltic_Main
+        String telemetryUrl = match.getTelemetryURL();
 
         assertThat(matchId).isEqualTo("b79c05af-bfb3-4744-b9d2-9ccdbe4c0746");
         assertThat(mapName).isEqualTo("Baltic_Main");
-        assertThat(includesSize).isEqualTo(3);
+        assertThat(telemetryUrl).isEqualTo(
+                "https://telemetry-cdn.pubg.com/bluehole-pubg/kakao/2024/10/18/16/29/2b177796-8d6e-11ef-aae5-8e2d0dc0c1b1-telemetry.json");
 
     }
